+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+### Added
+
+- New flag `--hcl-provider-block` to be able to opt out of the `provider "" {}` on HCL
+  ([Issue #250](https://github.com/cycloidio/terracognita/issues/250))
+
 ## [0.7.3] _2021-09-23_
 
 ### Changed